Directly bound the Consumer Lambda to the SQS
I had a loose, wildcard Event Pattern previously, which was bound to get me into trouble. I don't think it was even tied to a specific Queue!

Lambda Handler function re-written to use the argv, rather than to invoke the sqs api.

Modified the Service Role also to allow for CloudWatch logging. This was necessary for debugging. Testing the Consumer with test events no longer works effectively because of the code no longer directly references the queue through boto3. Hence - you have to let the consumer rip, print to stdout (which results in logging), and then debug from the logs.